Esercizione

+++ src/main/java/com/generation/sitoillogico/controller/Cla.java
@@ -1,29 +1,41 @@
 package com.generation.sitoillogico.controller;
 
+import com.generation.sitoillogico.model.dao.AltroDao;
 import com.generation.sitoillogico.model.dao.QualcosaDao;
+import com.generation.sitoillogico.model.dao.UtenteDao;
+import com.generation.sitoillogico.model.entities.Altro;
 import com.generation.sitoillogico.model.entities.Qualcosa;
+import com.generation.sitoillogico.model.entities.Utente;
 import com.generation.sitoillogico.model.entities.enums.Enumeratore;
+import org.apache.commons.codec.digest.DigestUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.support.SessionStatus;
 
 import java.time.LocalDate;
 import java.util.List;
-import java.util.stream.Collectors;
 
-
+@SessionAttributes({"utente"})
 @Controller
 public class Cla
 {
     @Autowired
     QualcosaDao qualcosaDao;
 
+    @Autowired
+    AltroDao altroDao;
+
+    @Autowired
+    UtenteDao utenteDao;
+
     @GetMapping("/qwe")
-    public String abc()
+    public String abc(@SessionAttribute(required = false) Utente utente)
     {
+        if (utente == null)
+            return "nonSeiLoggato";
+
         return "abc";
     }
 
@@ -54,8 +66,11 @@ public class Cla
     }
 
     @GetMapping("/uio")
-    public String uio(Model model)
+    public String uio(Model model, @SessionAttribute(required = false) Utente utente)
     {
+        if (utente == null)
+            return "nonSeiLoggato";
+
         List<Qualcosa> tuttiQualcosa = qualcosaDao.findAll();
         List<Qualcosa> qualcosaFiltrati = tuttiQualcosa.stream().
                 filter(qualcosa -> qualcosa.getC() == Enumeratore.BBB).
@@ -66,16 +81,92 @@ public class Cla
     }
 
     @GetMapping("/pas")
-    public String pas()
+    public String pas(@SessionAttribute(required = false) Utente utente)
     {
+        if (utente == null)
+            return "nonSeiLoggato";
+
         return "lmn";
     }
 
     @GetMapping("/DESTRUCTION")
-    public String destruction()
+    public String destruction(@SessionAttribute(required = false) Utente utente)
     {
+        if (utente == null)
+            return "nonSeiLoggato";
+
         qualcosaDao.deleteAll();
 
         return "desolazione";
+    }
+
+    @GetMapping("/dfg")
+    public String dfg(Model model, Long a)
+    {
+        model.addAttribute("idQualcosa", qualcosaDao.findById(a).orElse(null));
+        return "opq";
+    }
+
+    @PostMapping("/dfg")
+    public String PostOfg
+            (
+                    @RequestParam Long a,
+                    @RequestParam String f
+            )
+    {
+        Qualcosa qualcosa = qualcosaDao.findById(a).orElse(null);
+
+        Altro altro = new Altro();
+        altro.setF(f);
+        altro.setQualcosa(qualcosa);
+
+        altroDao.save(altro);
+
+        return "redirect:/dfg?a=" + a;
+    }
+
+    @PostMapping("/login")
+    public String login(Model model, @RequestParam String username, @RequestParam String password)
+    {
+        String passwordHashata = DigestUtils.md5Hex(password);
+        Utente utente = utenteDao.findByUsernameAndPassword(username, passwordHashata);
+
+        if (utente == null)
+            return "utenteNotFound";
+
+        model.addAttribute("utente", utente);
+
+        return "loginCorretto";
+    }
+
+    @GetMapping("/registrati")
+    public String apriPaginaRegistrazione()
+    {
+        return "registrazione";
+    }
+
+    @PostMapping("/registrati")
+    public String registrati(@RequestParam String username, @RequestParam String password)
+    {
+        Utente utente = new Utente();
+        utente.setUsername(username);
+
+        String passwordHashata = DigestUtils.md5Hex(password);
+        utente.setPassword(passwordHashata);
+
+        utenteDao.save(utente);
+
+        return "redirect:/";
+    }
+
+    @GetMapping("/sloggati")
+    public String sloggati(SessionStatus sessionStatus, @SessionAttribute(required = false) Utente utente)
+    {
+        if (utente == null)
+            return "nonSeiLoggato";
+
+        sessionStatus.setComplete();
+
+        return "redirect:/";
     }
 }
